+++ src/task.h
@@ -6,7 +6,7 @@ namespace zombie {
 	class Task {
 	public:
 		// The current size of the game windows.
-		static int width, height;
+		static float width, height;
 
 		virtual ~Task() {
 		}
